rules.in: dont need to care about BUILD_SYS
The -X argument to dh_fixperms says to eXclude linuxcnc_module_helper
when it's fixing the permissions on all our files.  But dh_fixperm won't
complain or fail if it doesn't find linuxcnc_module_helper, so we don't
need to do this substitution trick to tell it what to expect.